pyproject.toml
[tool.poetry]
name = "camera-trap-tk"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.2.0"
tksheet = "^5.0.25"
pyinstaller = "^4.3"
requests = "^2.25.1"
boto3 = "^1.17.79"
ttkwidgets = "^0.11.0"
flake8 = "^4.0.1"
bson = "^0.5.10"
imageio = "^2.20.0"
imageio-ffmpeg = "^0.4.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
